Correct buildability after moving some mock-classes
Related tasks APPLINK-30588 APPLINK-30974
