+++ src/components/application_manager/test/usage_statistics_test.cc
@@ -34,7 +34,7 @@
 #include <memory>
 #include "gmock/gmock.h"
 #include "smart_objects/enum_schema_item.h"
-#include "policy/mock_statistics_manager.h"
+#include "policy/usage_statistics/mock_statistics_manager.h"
 #include "mock_app_stopwatch.h"
 #include "utils/make_shared.h"
 #include "utils/shared_ptr.h"
